2-c-1 documentation added
Added docstrings so that if the code is adapted by others in the future, it is easier for them to follow.
Also found some of the one line functions were unnecessary as they did not improve clarity (it was very easy to see what they did without giving them an informative name).
I therefore 'unfunctioned' these.



Signed-off-by: EmmaWoodONS <Emma.Wood@ons.gov.uk>