finalise 15-1-2 update script
Rob added area of Protected Areas to the data, so I have adapted the script to take account of this.
I have also adapted the architecture to match that of other indicators (compile_tables.R etc).
We now use Series in combo with Units, so I also adapted this bit.

Signed-off-by: EmmaWoodONS <Emma.Wood@ons.gov.uk>