changed series to match SDMX
We are thinking of moving to SDMX soon, so I made the series compatible (though I think perhaps slightly less user firendly). I have left the ad hoc as it is as, while it doesn;t match up brilliantly, I think it makes more sense to not split freshwater and terrestrial.

Signed-off-by: EmmaWoodONS <Emma.Wood@ons.gov.uk>