bug fixes
I initially ran this using a csv that had 0s instead of -s and only 2dp.
Re-ran with more dp and spotted that code broke when -s where not replaced with 0s.
This fix assumes that - really does mean 0 (checked with Robert and he connfirmed this was the case at least in this year)

Signed-off-by: EmmaWoodONS <Emma.Wood@ons.gov.uk>