bugfix following review
My laptop was reading in country as a factor rather than a character which led to differences
in how the script ran. Country in one df was in uppercase while in the others it
was in title case so when the dfs were joined, the country data were lost.

All Country columns in all dfs should now be in the same case, regardless of
whether they are read in as factors or character strings.

Signed-off-by: EmmaWoodONS <Emma.Wood@ons.gov.uk>