QA added, and sex removed
Sex is currently in the live data, but on reflection doesn't make sense
for this indicator (why does it matter how many female health workers there
are per 10 000 women, or how many female health workers there are per 10 000 people).
The fact that there are more female health workers in general is an interesting
point but not really relevant in the context of rate.